Agent: satisfy clippy on current stable, on both platforms
First CI run flagged 16 lints the local 1.93 toolchain does not carry: as_chunks
for the constant-size chunk loops (synthetic capture, clipboard PNG expansion,
software encoder, privacy mosaic, cursor masks, probe fixtures), is_multiple_of
for the watchdog tick, a while-let congestion loop, and dead Windows-only items
(the unused anyhow import, a redundant usize cast, the never-read out_width /
out_height on the Media Foundation processor). account() and Pacer::current_scale
are macOS-only, so they are gated rather than deleted.

+++ crates/agent/src/platform/windows.rs
@@ -2,7 +2,7 @@
 //! indicator window, `SendSAS`, and helpers for launching a process in the active session.
 
 use crate::approval::{ApprovalOutcome, IndicatorHandle};
-use anyhow::{anyhow, bail, Context, Result};
+use anyhow::{bail, Context, Result};
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::{mpsc, Arc};
 use std::time::Duration;
@@ -482,7 +482,7 @@ unsafe extern "system" fn chat_wndproc(
     }
     match msg {
         WM_COMMAND => {
-            let id = (wparam.0 & 0xffff) as usize;
+            let id = wparam.0 & 0xffff;
             if !shared_ptr.is_null() {
                 let shared = &*shared_ptr;
                 match id {
